i2c-core: changed arguments to some functions.
__i2c_start_xfer are now passed the adapter and the shared info from the client ordering the
transmission. Also i2c_master_tmo is passed the adapter instead of the client.

+++ include/sys/mem.h
@@ -64,7 +64,7 @@ struct heap_node
          * 
          * This describes the size of the node.
          */
-        unsigned short size;
+        size_t size;
 } __PACK__;
 typedef struct heap_node HEAPNODE;
